refactor task inheritance
