+++ src/task-component.js
@@ -1,11 +1,12 @@
-import {createElement} from "./util";
+import {BaseComponent} from './base-component';
 
-export class Component {
+export class TaskComponent extends BaseComponent {
   constructor(data) {
-    if (new.target === Component) {
+    if (new.target === TaskComponent) {
       throw new Error(`Can't instantiate BaseComponent, only concrete one.`);
     }
 
+    super();
     this._title = data.title;
     this._tags = data.tags;
     this._picture = data.picture;
@@ -14,9 +15,6 @@ export class Component {
     this._repeatingDays = data.repeatingDays;
     this._isFavorite = data.isFavorite;
     this._isDone = data.isDone;
-
-    this._element = null;
-    this._state = {};
   }
 
   _renderControls() {
@@ -92,30 +90,4 @@ export class Component {
   _isRepeated() {
     return Object.values(this._repeatingDays).some((it) => it);
   }
-
-  get element() {
-    return this._element;
-  }
-
-  get template() {
-    throw new Error(`You have to define template.`);
-  }
-
-  bind() {}
-
-  unbind() {}
-
-  render() {
-    this._element = createElement(this.template);
-    this.bind();
-    return this._element;
-  }
-
-  unrender() {
-    this.unbind();
-    this._element.remove();
-    this._element = null;
-  }
-
-  update() {}
 }
